Update package names, change download URL/version
Change 'ruby1.9.1' and 'ruby1.9.1-dev' for the metapackages that also have installation candidates for Ubuntu 16 & 18. This adds support for Ubuntu 16 & 18. Also add 'unzip' package as a dependency.

Change download URL and version to onetimesecret fork. This was necessary because the Gemfile/Gemfile.lock in the official versions referenced a broken version of the gem (yajl-ruby). See https://github.com/onetimesecret/onetimesecret/issues/81

A bit of linting.